# dennic_api_gateway/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Dennic Project",
        "contact": {},
        "version": "1.7"
    },
    "host": "swag.dennic.uz",
    "paths": {
...
        "/v1/doctor": {
            "get": {
                "description": "ListDoctors - Api for list doctor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor"
                ],
                "summary": "ListDoctors",
                "parameters": [
                    {
                        "type": "string",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "value",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "first_name",
                            "last_name",
                            "gender",
                            "phone_number",
                            "email",
                            "address",
                            "city",
                            "country",
                            "biography"
                        ],
                        "type": "string",
                        "description": "search",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorAndDoctorHours"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "put": {
                "description": "UpdateDoctor - Api for update doctor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor"
                ],
                "summary": "UpdateDoctor",
                "parameters": [
                    {
                        "description": "UpdateDoctorReq",
                        "name": "UpdateDoctorReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "post": {
                "description": "CreateDoctor - Api for crete doctor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor"
                ],
                "summary": "CreateDoctor",
                "parameters": [
                    {
                        "description": "DoctorReq",
                        "name": "DoctorReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            },
            "delete": {
                "description": "DeleteDoctor - Api for delete doctor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor"
                ],
                "summary": "DeleteDoctor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.StatusRes"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
...
        "/v1/doctor/get": {
            "get": {
                "description": "GetDoctor - Api for get doctor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor"
                ],
                "summary": "GetDoctor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.DoctorAndDoctorHours"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
        "/v1/doctor/spec": {
            "get": {
                "description": "ListDoctorsBySpecializationId - Api for list doctors by specialization id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Doctor"
                ],
                "summary": "ListDoctorsBySpecializationId",
                "parameters": [
                    {
                        "type": "string",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "value",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "specialization_id",
                        "name": "specialization_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model_healthcare_service.ListDoctors"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model_common.StandardErrorModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model_common.ResponseError": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model_common.StandardErrorModel": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/model_common.ResponseError"
                }
            }
        },
...
        "model_healthcare_service.DoctorAndDoctorHours": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "day_of_week": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "department_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "end_work_date": {
                    "type": "string"
                },
                "finish_time": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "room_number": {
                    "type": "integer"
                },
                "salary": {
                    "type": "number"
                },
                "specializations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model_healthcare_service.DoctorSpec"
                    }
                },
                "start_time": {
                    "type": "string"
                },
                "start_work_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "work_years": {
                    "type": "integer"
                }
            }
        },
...
        "model_healthcare_service.DoctorRes": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bio": {
                    "type": "string"
                },
                "birth_date": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "department_id": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "end_work_date": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "order": {
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "phone_number": {
                    "type": "string"
                },
                "room_number": {
                    "type": "integer"
                },
                "salary": {
                    "type": "number"
                },
                "start_work_date": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "work_years": {
                    "type": "integer"
                }
            }
        },
...
        "model_healthcare_service.DoctorSpec": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model_healthcare_service.DoctorUpdateReq": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Addres"
                },
                "bio": {
                    "type": "string",
                    "example": "Biography"
                },
                "birth_date": {
                    "type": "string",
                    "example": "2012-12-12"
                },
                "city": {
                    "type": "string",
                    "example": "City"
                },
                "country": {
                    "type": "string",
                    "example": "Country"
                },
                "department_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174001"
                },
                "email": {
                    "type": "string",
                    "example": "email@gmail.com"
                },
                "end-work-date": {
                    "type": "string",
                    "example": "2022-12-12"
                },
                "first_name": {
                    "type": "string",
                    "example": "First Name"
                },
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614274001"
                },
                "image_url": {
                    "type": "string",
                    "example": "http://example.com/image.png"
                },
                "last_name": {
                    "type": "string",
                    "example": "Last Name"
                },
                "password": {
                    "type": "string",
                    "example": "password"
                },
                "phone_number": {
                    "type": "string",
                    "example": "+998901234567"
                },
                "room_number": {
                    "type": "integer",
                    "example": 1
                },
                "salary": {
                    "type": "number",
                    "example": 10
                },
                "start_work_date": {
                    "type": "string",
                    "example": "2012-12-12"
                },
                "work_years": {
                    "type": "integer",
                    "example": 4
                }
            }
        },
...
        "model_healthcare_service.ListDoctors": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "doctors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model_healthcare_service.DoctorRes"
                    }
                }
            }
        },
...
        "models.StatusRes": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}